Unify all hamburger menus to use same style
