+++ app/src/main/java/org/sil/storyproducer/controller/SplashScreenActivity.kt
@@ -29,4 +29,8 @@ class SplashScreenActivity : BaseActivity() {
 
     }
 
+    override fun setupStoryListTabPages() {
+        // nothing to implement
+    }
+
 }
