+++ app/src/main/res/layout/phase_frame.xml
@@ -8,7 +8,6 @@
     android:focusable="true"
     android:focusableInTouchMode="true">
 
-
     <com.google.android.material.appbar.AppBarLayout
         android:layout_height="wrap_content"
         android:layout_width="match_parent">
